[slip ci] Update options on separate page for output. Frontend
