+++ components/restreamer/package.json
@@ -32,6 +32,7 @@
     "cross-env": "^5.2",
     "css-loader": "^2.1",
     "graphql-tag": "^2.11",
+    "html-webpack-plugin": "^4.5.1",
     "lodash.clonedeep": "^4.5.0",
     "mini-css-extract-plugin": "^0.6",
     "minify-html-webpack-plugin": "^1.1",
