--- components/restreamer/static/index.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf8" />
-        <meta name="robots" content="noindex, nofollow" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-
-        <title>Ephyr re-streamer</title>
-
-        <link rel="icon" type="image/png" sizes="48x48" href="favicon.ico" />
-        <link
-            rel="icon"
-            type="image/png"
-            sizes="32x32"
-            href="favicon-32x32.png"
-        />
-        <link
-            rel="icon"
-            type="image/png"
-            sizes="16x16"
-            href="favicon-16x16.png"
-        />
-
-        <link
-            rel="stylesheet"
-            href="https://pro.fontawesome.com/releases/v5.14.0/css/all.css"
-        />
-        <link rel="stylesheet" href="bundle.css" />
-    </head>
-
-    <body>
-        <script
-            defer
-            src="https://use.fontawesome.com/releases/v5.14.0/js/all.js"
-        ></script>
-        <script src="bundle.js"></script>
-    </body>
-</html>
